Rename FhirMappingAdapterDSTU2
